feat(mobile): archived threads screen + archive/unarchive actions
