+++ uxnanmobile/l10n/app_en.arb
@@ -81,7 +81,12 @@
   "threadsEmptyBody": "Threads from this PC will appear here. Pull down to refresh.",
   "threadActionRename": "Rename",
   "threadActionCopyId": "Copy thread ID",
+  "threadActionArchive": "Archive",
+  "threadActionUnarchive": "Unarchive",
   "threadActionDelete": "Delete",
+  "archivedTitle": "Archived",
+  "archivedEmpty": "No archived threads",
+  "archivedEmptyBody": "Threads you archive are hidden here, not deleted. Long-press one to unarchive it.",
   "threadRenameTitle": "Rename thread",
   "threadRenameHint": "Thread title",
   "threadIdCopied": "Thread ID copied",
