+++ uxnanmobile/l10n/app_es.arb
@@ -66,7 +66,12 @@
   "threadsEmptyBody": "Las conversaciones de esta PC aparecerán aquí. Desliza para actualizar.",
   "threadActionRename": "Renombrar",
   "threadActionCopyId": "Copiar ID del hilo",
+  "threadActionArchive": "Archivar",
+  "threadActionUnarchive": "Desarchivar",
   "threadActionDelete": "Eliminar",
+  "archivedTitle": "Archivados",
+  "archivedEmpty": "No hay conversaciones archivadas",
+  "archivedEmptyBody": "Las conversaciones que archivas se ocultan aquí, no se eliminan. Mantén pulsada una para desarchivarla.",
   "threadRenameTitle": "Renombrar hilo",
   "threadRenameHint": "Título del hilo",
   "threadIdCopied": "ID del hilo copiado",
