+++ uxnanmobile/lib/presentation/screens/threads/threads_screen.dart
@@ -1,11 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
-import 'package:intl/intl.dart';
 import 'package:uxnan/domain/entities/thread.dart';
 import 'package:uxnan/domain/entities/trusted_device.dart';
 import 'package:uxnan/domain/enums/agent_id.dart';
@@ -15,9 +13,9 @@ import 'package:uxnan/l10n/app_localizations.dart';
 import 'package:uxnan/presentation/providers/application_providers.dart';
 import 'package:uxnan/presentation/router/app_router.dart';
 import 'package:uxnan/presentation/screens/threads/new_conversation_sheet.dart';
+import 'package:uxnan/presentation/screens/threads/thread_tile.dart';
 import 'package:uxnan/presentation/theme/colors.dart';
 import 'package:uxnan/presentation/theme/spacing.dart';
-import 'package:uxnan/presentation/widgets/agent_logo_chip.dart';
 import 'package:uxnan/presentation/widgets/agent_visuals.dart';
 
 /// The threads of a connected PC (spec 02a §5.4.2). Lists the active bridge's
@@ -77,9 +75,11 @@ class _ThreadsScreenState extends ConsumerState<ThreadsScreen> {
   @override
   Widget build(BuildContext context) {
     final allThreads = ref.watch(threadsProvider).value ?? const <Thread>[];
-    // Scope to the selected PC. Legacy threads with no device tag are still
-    // shown (they get tagged on the next connected refresh); demo data is gone.
+    // Scope to the selected PC and hide archived threads (those live on the
+    // Archived screen). Legacy threads with no device tag are still shown (they
+    // get tagged on the next connected refresh); demo data is gone.
     final threads = allThreads
+        .where((t) => t.status != ThreadStatus.archived)
         .where((t) => t.deviceId == null || t.deviceId == widget.deviceId)
         .toList();
     final devices = ref.watch(trustedDevicesProvider).value ?? const [];
@@ -119,6 +119,13 @@ class _ThreadsScreenState extends ConsumerState<ThreadsScreen> {
               snap: true,
               title: Text(_title(devices), overflow: TextOverflow.ellipsis),
               actions: [
+                IconButton(
+                  tooltip: l10n.archivedTitle,
+                  icon: const Icon(Icons.archive_outlined),
+                  onPressed: () => context.push(
+                    AppRoutes.deviceArchived(widget.deviceId),
+                  ),
+                ),
                 _ConnectionDot(phase: phase),
                 const SizedBox(width: UxnanSpacing.lg),
               ],
@@ -149,7 +156,7 @@ class _ThreadsScreenState extends ConsumerState<ThreadsScreen> {
                   separatorBuilder: (_, __) =>
                       const SizedBox(height: UxnanSpacing.md),
                   itemBuilder: (context, index) =>
-                      _ThreadTile(thread: visible[index]),
+                      ThreadTile(thread: visible[index]),
                 ),
               ),
           ],
@@ -241,267 +248,6 @@ class _AgentChipAvatar extends StatelessWidget {
   }
 }
 
-/// A per-thread action chosen from the long-press menu.
-enum _ThreadAction { rename, copyId, delete }
-
-class _ThreadTile extends ConsumerWidget {
-  const _ThreadTile({required this.thread});
-  final Thread thread;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final colors = Theme.of(context).colorScheme;
-    final textTheme = Theme.of(context).textTheme;
-    final agent = AgentIdParsing.fromWireId(thread.agentId);
-
-    return Material(
-      color: colors.surfaceContainerHighest,
-      borderRadius: const BorderRadius.all(UxnanRadius.lg),
-      child: InkWell(
-        borderRadius: const BorderRadius.all(UxnanRadius.lg),
-        onTap: () => context.push(AppRoutes.conversation(thread.id)),
-        onLongPress: () => _showThreadActions(context, ref, thread),
-        child: Padding(
-          padding: const EdgeInsets.all(UxnanSpacing.md),
-          child: Row(
-            children: [
-              _AgentAvatar(agent: agent),
-              const SizedBox(width: UxnanSpacing.md),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            thread.title,
-                            style: textTheme.titleSmall,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                        if (thread.lastActivity != null) ...[
-                          const SizedBox(width: UxnanSpacing.sm),
-                          Text(
-                            _relativeTime(thread.lastActivity!),
-                            style: textTheme.bodySmall?.copyWith(
-                              color: colors.onSurfaceVariant,
-                            ),
-                          ),
-                        ],
-                      ],
-                    ),
-                    const SizedBox(height: UxnanSpacing.xs),
-                    Row(
-                      children: [
-                        _StatusDot(status: thread.status),
-                        const SizedBox(width: UxnanSpacing.xs),
-                        Flexible(
-                          child: Text(
-                            _subtitle(),
-                            style: textTheme.bodySmall?.copyWith(
-                              color: colors.onSurfaceVariant,
-                            ),
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  String _subtitle() {
-    final agent = AgentVisuals.labelFor(
-      AgentIdParsing.fromWireId(thread.agentId),
-    );
-    final dir = thread.cwd?.split(RegExp(r'[\\/]')).last;
-    return dir == null ? agent : '$agent · $dir';
-  }
-}
-
-/// Shows the per-thread actions sheet (rename / copy id / delete) on long-press.
-Future<void> _showThreadActions(
-  BuildContext context,
-  WidgetRef ref,
-  Thread thread,
-) async {
-  final l10n = AppLocalizations.of(context);
-  final colors = Theme.of(context).colorScheme;
-  final action = await showModalBottomSheet<_ThreadAction>(
-    context: context,
-    showDragHandle: true,
-    builder: (context) => SafeArea(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          ListTile(
-            title: Text(thread.title, overflow: TextOverflow.ellipsis),
-            subtitle: Text(
-              thread.id,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-          ),
-          const Divider(height: 1),
-          ListTile(
-            leading: const Icon(Icons.edit_outlined),
-            title: Text(l10n.threadActionRename),
-            onTap: () => Navigator.pop(context, _ThreadAction.rename),
-          ),
-          ListTile(
-            leading: const Icon(Icons.content_copy_outlined),
-            title: Text(l10n.threadActionCopyId),
-            onTap: () => Navigator.pop(context, _ThreadAction.copyId),
-          ),
-          ListTile(
-            leading: Icon(Icons.delete_outline, color: colors.error),
-            title: Text(
-              l10n.threadActionDelete,
-              style: TextStyle(color: colors.error),
-            ),
-            onTap: () => Navigator.pop(context, _ThreadAction.delete),
-          ),
-        ],
-      ),
-    ),
-  );
-  if (action == null || !context.mounted) return;
-  switch (action) {
-    case _ThreadAction.rename:
-      await _promptRenameThread(context, ref, thread);
-    case _ThreadAction.copyId:
-      await Clipboard.setData(ClipboardData(text: thread.id));
-      if (context.mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text(l10n.threadIdCopied)),
-        );
-      }
-    case _ThreadAction.delete:
-      await _confirmDeleteThread(context, ref, thread);
-  }
-}
-
-/// Prompts for a new title and renames the thread via the thread manager.
-Future<void> _promptRenameThread(
-  BuildContext context,
-  WidgetRef ref,
-  Thread thread,
-) async {
-  final l10n = AppLocalizations.of(context);
-  final controller = TextEditingController(text: thread.title);
-  final newTitle = await showDialog<String>(
-    context: context,
-    builder: (context) => AlertDialog(
-      title: Text(l10n.threadRenameTitle),
-      content: TextField(
-        controller: controller,
-        autofocus: true,
-        textInputAction: TextInputAction.done,
-        decoration: InputDecoration(labelText: l10n.threadRenameHint),
-        onSubmitted: (value) => Navigator.pop(context, value),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: Text(l10n.actionCancel),
-        ),
-        FilledButton(
-          onPressed: () => Navigator.pop(context, controller.text),
-          child: Text(l10n.actionSave),
-        ),
-      ],
-    ),
-  );
-  controller.dispose();
-  final trimmed = newTitle?.trim() ?? '';
-  if (trimmed.isEmpty || trimmed == thread.title) return;
-  await ref.read(threadManagerProvider).renameThread(thread.id, trimmed);
-}
-
-/// Confirms and deletes the thread via the thread manager.
-Future<void> _confirmDeleteThread(
-  BuildContext context,
-  WidgetRef ref,
-  Thread thread,
-) async {
-  final l10n = AppLocalizations.of(context);
-  final colors = Theme.of(context).colorScheme;
-  final confirmed = await showDialog<bool>(
-    context: context,
-    builder: (context) => AlertDialog(
-      title: Text(l10n.threadDeleteTitle),
-      content: Text(l10n.threadDeleteBody),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context, false),
-          child: Text(l10n.actionCancel),
-        ),
-        FilledButton(
-          style: FilledButton.styleFrom(backgroundColor: colors.error),
-          onPressed: () => Navigator.pop(context, true),
-          child: Text(l10n.threadDeleteConfirm),
-        ),
-      ],
-    ),
-  );
-  if (confirmed != true) return;
-  await ref.read(threadManagerProvider).deleteThread(thread.id);
-}
-
-class _AgentAvatar extends StatelessWidget {
-  const _AgentAvatar({required this.agent});
-  final AgentId agent;
-
-  @override
-  Widget build(BuildContext context) {
-    final logo = AgentVisuals.logoFor(agent);
-    if (logo != null) return AgentLogoChip(asset: logo, size: 44);
-
-    final colors = Theme.of(context).colorScheme;
-    return Container(
-      width: 44,
-      height: 44,
-      decoration: BoxDecoration(
-        color: colors.surfaceContainerHigh,
-        borderRadius: const BorderRadius.all(UxnanRadius.lg),
-        border: Border.all(color: colors.outline),
-      ),
-      child: Icon(
-        Icons.smart_toy_outlined,
-        size: 22,
-        color: AgentVisuals.colorFor(agent),
-      ),
-    );
-  }
-}
-
-class _StatusDot extends StatelessWidget {
-  const _StatusDot({required this.status});
-  final ThreadStatus status;
-
-  @override
-  Widget build(BuildContext context) {
-    final color = switch (status) {
-      ThreadStatus.active => UxnanColors.connected,
-      ThreadStatus.syncing => UxnanColors.syncing,
-      ThreadStatus.error => UxnanColors.error,
-      ThreadStatus.archived => UxnanColors.onSurfaceMuted,
-    };
-    return Container(
-      width: 6,
-      height: 6,
-      decoration: BoxDecoration(color: color, shape: BoxShape.circle),
-    );
-  }
-}
-
 class _ConnectionDot extends StatelessWidget {
   const _ConnectionDot({required this.phase});
   final ConnectionPhase phase;
@@ -562,13 +308,4 @@ class _EmptyThreads extends StatelessWidget {
       ),
     );
   }
-}
-
-String _relativeTime(DateTime time) {
-  final now = DateTime.now();
-  final isSameDay =
-      now.year == time.year && now.month == time.month && now.day == time.day;
-  return isSameDay
-      ? DateFormat.Hm().format(time)
-      : DateFormat.MMMd().format(time);
 }
